Add another web based fretboard tool
I see that there is some overlap with the tools already on the list. A feature I haven't seen in the other tools is the highlighting of the notes in the triads that can be build from a scale. So maybe you find it a useful addition. Disclaimer: I created this.